% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findChromPeaks.R
\name{findChromPeaks_CWT}
\alias{findChromPeaks_CWT}
\title{findChromPeaks_CWT}
\usage{
findChromPeaks_CWT(
  int,
  rt,
  peakwidth = c(5, 20),
  snthresh = 10,
  minPs = 3,
  noise = 100,
  estimateNoise = TRUE,
  extendLengthMSW = TRUE,
  r2thresh = 0.6,
  csthresh = 0.2
)
}
\arguments{
\item{int}{\code{numeric}, intensity of chromatogram.}

\item{rt}{\code{numeric}, retention time of chromatogram.}

\item{peakwidth}{\code{numeric(2)} with the lower and upper boun of the expected peak width.}

\item{snthresh}{\code{numeric(1)} defining the signal to noise ratio cutoff.}

\item{minPs}{\code{integer(1)}, the ROI region requires a minimum of minPs of signals greater than the noise.}

\item{noise}{\code{numeric(1)}, noise of chromatogram.}

\item{estimateNoise}{\code{logical(1)}, whether to estimate noise}

\item{extendLengthMSW}{\code{logical(1)} please see: \link[xcms:findChromPeaks-centWave]{xcms::centWave}}

\item{r2thresh}{\code{numeric(1)} threshold of peak shape.}

\item{csthresh}{\code{numeric(1)} threshold of cs.}
}
\value{
\code{matrix}
}
\description{
Find chromatographic peaks using CentWave.
}
\examples{
rtime <- seq(0, 1000, 0.5)
peak1 <- gaussian_peak(rtime, peakRt = 240, peakWidth = 20, peakHeight = 10000)
peak2 <- tailing_peak(rtime, peakRt = 480, peakWidth = 25, peakHeight = 1300, tau = 10)
peak3 <- leading_peak(rtime, peakRt = 720, peakWidth = 18, peakHeight = 5000, tau = -8)
combined <- peak1 + peak2 + peak3
baseline <- 100 + 1 * rtime + 100 * sin(rtime/100)
final_chrom <- combined + baseline  + abs(rnorm(length(rtime), mean = 0, sd = 25))
plot(rtime, final_chrom, type = "l")
findChromPeaks_CWT(int = final_chrom, rt = rtime, peakwidth = c(5, 30), snthresh = 2, noise = 100)
xcms::peaksWithCentWave(int = final_chrom, rt = rtime, peakwidth = c(5, 30), snthresh = 2, noise = 100)
}
