% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGauss.R
\name{fitGauss}
\alias{fitGauss}
\title{fitGauss}
\usage{
fitGauss(int)
}
\arguments{
\item{int}{\code{numeric}, intensity of a peak.}
}
\value{
\code{numeric(1)}, r2 value
}
\description{
Assessing whether a peak is a Gaussian peak
}
\examples{
fitGauss(int)
}
