% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateChromPeaks.R
\name{gaussian_peak}
\alias{gaussian_peak}
\alias{tailing_peak}
\alias{leading_peak}
\title{Simulate chromatographic peaks}
\usage{
gaussian_peak(rtime, peakRt, peakWidth, peakHeight)

tailing_peak(rtime, peakRt, peakWidth, peakHeight, tau = 10)

leading_peak(rtime, peakRt, peakWidth, peakHeight, tau = -10)
}
\arguments{
\item{rtime}{\code{numeric()}, retention time vector of chromatogram}

\item{peakRt}{\code{numeric(1)}, retention time of peak apex}

\item{peakWidth}{\code{numeric(1)}, peak width}

\item{peakHeight}{\code{numeric(1)}, peak height}

\item{tau}{\code{numeric(1)}, the degree of tailing or leading of chromatographic peak;
for tailing peas, greater than 0; for leading less than 0}
}
\value{
\code{numeric()}, intensity vector of chromatogram
}
\description{
Simulate three types of chromatographic peaks:
gaussian symmetrical peak, tailing peak and leading peak
}
\examples{
rtime <- seq(0, 1000, 0.5)
peak1 <- gaussian_peak(rtime, peakRt = 240, peakWidth = 20, peakHeight = 10000)
peak2 <- tailing_peak(rtime, peakRt = 480, peakWidth = 25, peakHeight = 1300, tau = 10)
peak3 <- leading_peak(rtime, peakRt = 720, peakWidth = 18, peakHeight = 5000, tau = -8)
combined <- peak1 + peak2 + peak3
baseline <- 100 + 1 * rtime + 100 * sin(rtime/100)
final_chrom <- combined + baseline  + abs(rnorm(length(rtime), mean = 0, sd = 25))
plot(rtime, final_chrom, type = "l")
}
